refactor(markdown): drop the compile package and hide the fold behind Compiler
The algebra and its driver were landing in morphir.langkit.markdown.compile.
A public sub-package makes a caller import two packages to parse and then
compile one document, and it freezes today's internal decomposition into the
published surface, so moving a type between roles later would break consumers.

Public types now live in morphir.langkit.markdown, and depth goes downward
instead of sideways. The traversal moves to morphir.langkit.markdown.internal
as a private[markdown] MarkdownFold, and Compiler re-exports its entry point:

    object Compiler:
      export morphir.langkit.markdown.internal.MarkdownFold.compile

Callers write Compiler.compile(document) and cannot name MarkdownFold, so the
walk can gain node kinds as intent 0021 widens the AST without any of that
reaching the published API. An export clause is public even when its target is
private[markdown], which is what makes the forwarder the API and the target
not. Confirmed by compiling a probe from an outside package and watching it
fail, then deleting it. This follows morphir.buildkit.internal, which the root
AGENTS.md names as the pattern for Kyo-based modules.

Adds morphir/langkit/markdown/CONTRIBUTING.md recording the convention: one
public package, internal for machinery, export for entry points, plus how to
check the boundary actually holds. Design rationale that had accumulated in
the Compiler scaladoc — why a Monoid cannot express the shape, why a visitor
was rejected — moves there, leaving the doc comment to say what the type is
and how to implement it.

52 tests pass on JVM, JS and Native; morphir-knowledge-okf still compiles.
