+++ morphir/langkit/markdown/src/morphir/langkit/markdown/internal/MarkdownFold.scala
@@ -1,4 +1,4 @@
-package morphir.langkit.markdown.compile
+package morphir.langkit.markdown.internal
 
 import kyo.*
 import morphir.langkit.markdown.*
@@ -7,11 +7,13 @@ import morphir.langkit.markdown.*
  * The single owner of Markdown AST traversal.
  *
  * Walks bottom-up: children compile first, then each node combines them into one `Out`. Every output format reuses this
- * walk by supplying a [[Compiler]] instead of writing its own.
+ * walk by supplying a [[morphir.langkit.markdown.Compiler]] instead of writing its own.
+ *
+ * Internal on purpose. Callers reach it as `Compiler.compile`, which exports the entry point below, so the traversal
+ * can gain node kinds as intent 0021 widens the AST without that being a change to the public surface.
  */
-object MarkdownCompiler:
+private[markdown] object MarkdownFold:
 
-  /** Compile a whole document with the given output format. */
   def compile[Out](document: Document)(using compiler: Compiler[Out]): Out =
     compiler.document(document.blocks.map(compileBlock))
 
@@ -22,4 +24,4 @@ object MarkdownCompiler:
       case Block.FencedCode(info, content, _) => compiler.fencedCode(info, content)
       case Block.UnorderedList(items, _)      => compiler.unorderedList(items.map(compiler.listItem))
       case Block.ThematicBreak(_)             => compiler.thematicBreak
-end MarkdownCompiler
+end MarkdownFold
